+++ src/model/Model_Checking.cpp
@@ -128,7 +128,7 @@ Model_Checking& Model_Checking::instance()
     return Singleton<Model_Checking>::instance();
 }
 
-bool Model_Checking::remove(int id)
+bool Model_Checking::remove(int64 id)
 {
     //TODO: remove all split at once
     //Model_Splittransaction::instance().remove(Model_Splittransaction::instance().find(Model_Splittransaction::TRANSID(id)));
@@ -145,7 +145,7 @@ bool Model_Checking::remove(int id)
     return this->remove(id, db_);
 }
 
-int Model_Checking::save(Data* r)
+int64 Model_Checking::save(Data* r)
 {
     wxSharedPtr<Data> oldData(instance().get_record(r->TRANSID));
     if (!oldData || (!oldData->equals(r) && oldData->DELETEDTIME.IsEmpty() && r->DELETEDTIME.IsEmpty()))
@@ -154,7 +154,7 @@ int Model_Checking::save(Data* r)
     return r->TRANSID;
 }
 
-int Model_Checking::save(std::vector<Data>& rows)
+int64 Model_Checking::save(std::vector<Data>& rows)
 {
     this->Savepoint();
     for (auto& r : rows)
@@ -168,7 +168,7 @@ int Model_Checking::save(std::vector<Data>& rows)
     return rows.size();
 }
 
-int Model_Checking::save(std::vector<Data*>& rows)
+int64 Model_Checking::save(std::vector<Data*>& rows)
 {
     this->Savepoint();
     for (auto& r : rows)
@@ -291,7 +291,7 @@ Model_Checking::STATUS_ID Model_Checking::status_id(const Data* r)
     return status_id(r->STATUS);
 }
 
-double Model_Checking::amount(const Data* r, int account_id)
+double Model_Checking::amount(const Data* r, int64 account_id)
 {
     double sum = 0;
     switch (type_id(r->TRANSCODE))
@@ -314,50 +314,50 @@ double Model_Checking::amount(const Data* r, int account_id)
     return sum;
 }
 
-double Model_Checking::amount(const Data&r, int account_id)
+double Model_Checking::amount(const Data&r, int64 account_id)
 {
     return amount(&r, account_id);
 }
 
-double Model_Checking::balance(const Data* r, int account_id)
+double Model_Checking::balance(const Data* r, int64 account_id)
 {
     if (Model_Checking::status_id(r->STATUS) == Model_Checking::STATUS_ID_VOID || !r->DELETEDTIME.IsEmpty()) return 0;
     return amount(r, account_id);
 }
 
-double Model_Checking::balance(const Data& r, int account_id)
+double Model_Checking::balance(const Data& r, int64 account_id)
 {
     return balance(&r, account_id);
 }
 
-double Model_Checking::withdrawal(const Data* r, int account_id)
+double Model_Checking::withdrawal(const Data* r, int64 account_id)
 {
     double bal = balance(r, account_id);
     return bal <= 0 ? -bal : 0;
 }
 
-double Model_Checking::withdrawal(const Data& r, int account_id)
+double Model_Checking::withdrawal(const Data& r, int64 account_id)
 {
     return withdrawal(&r, account_id);
 }
 
-double Model_Checking::deposit(const Data* r, int account_id)
+double Model_Checking::deposit(const Data* r, int64 account_id)
 {
     double bal = balance(r, account_id);
     return bal > 0 ? bal : 0;
 }
 
-double Model_Checking::deposit(const Data& r, int account_id)
+double Model_Checking::deposit(const Data& r, int64 account_id)
 {
     return deposit(&r, account_id);
 }
 
-double Model_Checking::reconciled(const Data* r, int account_id)
+double Model_Checking::reconciled(const Data* r, int64 account_id)
 {
     return (Model_Checking::status_id(r->STATUS) == Model_Checking::STATUS_ID_RECONCILED) ? balance(r, account_id) : 0;
 }
 
-double Model_Checking::reconciled(const Data& r, int account_id)
+double Model_Checking::reconciled(const Data& r, int64 account_id)
 {
     return reconciled(&r, account_id);
 }
@@ -452,8 +452,8 @@ Model_Checking::Full_Data::Full_Data(const Data& r) : Data(r), BALANCE(0), AMOUN
 }
 
 Model_Checking::Full_Data::Full_Data(const Data& r
-    , const std::map<int /*trans id*/, Model_Splittransaction::Data_Set /*split trans*/ > & splits
-    , const std::map<int /*trans id*/, Model_Taglink::Data_Set /*split trans*/ >& tags)
+    , const std::map<int64 /*trans id*/, Model_Splittransaction::Data_Set /*split trans*/ > & splits
+    , const std::map<int64 /*trans id*/, Model_Taglink::Data_Set /*split trans*/ >& tags)
     : Data(r), BALANCE(0), AMOUNT(0)
 {
     const auto it = splits.find(this->id());
@@ -502,7 +502,7 @@ Model_Checking::Full_Data::~Full_Data()
 {
 }
 
-wxString Model_Checking::Full_Data::real_payee_name(int account_id) const
+wxString Model_Checking::Full_Data::real_payee_name(int64 account_id) const
 {
     if (TYPE_ID_TRANSFER == type_id(this->TRANSCODE))
     {
@@ -515,7 +515,7 @@ wxString Model_Checking::Full_Data::real_payee_name(int account_id) const
     return this->PAYEENAME;
 }
 
-const wxString Model_Checking::Full_Data::get_currency_code(int account_id) const
+const wxString Model_Checking::Full_Data::get_currency_code(int64 account_id) const
 {
     if (TYPE_ID_TRANSFER == type_id(this->TRANSCODE))
     {
@@ -525,13 +525,13 @@ const wxString Model_Checking::Full_Data::get_currency_code(int account_id) cons
             account_id = this->TOACCOUNTID;
     }
     Model_Account::Data* acc = Model_Account::instance().get(account_id);
-    int currency_id = acc ? acc->CURRENCYID: -1;
+    int64 currency_id = acc ? acc->CURRENCYID: -1;
     Model_Currency::Data* curr = Model_Currency::instance().get(currency_id);
 
     return curr ? curr->CURRENCY_SYMBOL : "";
 }
 
-const wxString Model_Checking::Full_Data::get_account_name(int account_id) const
+const wxString Model_Checking::Full_Data::get_account_name(int64 account_id) const
 {
     if (TYPE_ID_TRANSFER == type_id(this->TRANSCODE))
     {
@@ -577,7 +577,7 @@ bool CompareUsedNotes(const std::tuple<int, wxString, wxString>& a, const std::t
     return false;
 }
 
-void Model_Checking::getFrequentUsedNotes(std::vector<wxString> &frequentNotes, int accountID)
+void Model_Checking::getFrequentUsedNotes(std::vector<wxString> &frequentNotes, int64 accountID)
 {
     frequentNotes.clear();
     size_t max = 20;
@@ -611,7 +611,7 @@ void Model_Checking::getFrequentUsedNotes(std::vector<wxString> &frequentNotes, 
     }
 }
 
-void Model_Checking::getEmptyTransaction(Data &data, int accountID)
+void Model_Checking::getEmptyTransaction(Data &data, int64 accountID)
 {
     data.TRANSID = -1;
     data.PAYEEID = -1;
@@ -755,7 +755,7 @@ bool Model_Checking::foreignTransactionAsTransfer(const Data& data)
     return foreignTransaction(data) && (data.TOACCOUNTID == Model_Translink::AS_TRANSFER);
 }
 
-void Model_Checking::updateTimestamp(int id)
+void Model_Checking::updateTimestamp(int64 id)
 {
     Data* r = instance().get(id);
     if (r && r->TRANSID == id) {
